requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
python-docx==1.1.0
Pillow>=10.1.0
pillow-heif>=0.13.0
Werkzeug==3.0.1
python-dotenv==1.0.0
twilio>=8.10.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
bleach==6.1.0
MarkupSafe==2.1.3
cloudinary==1.41.0
